Solution for part 2 of 7th of december

+++ 7December/solution.h
@@ -2,5 +2,5 @@
 #define SOLUTION7_H
 #include <bits/stdc++.h>
 uint64_t solution7();
-int second_solution7();
+uint64_t second_solution7();
 #endif
